sentiment_analysis/main/pyproject.toml
[tool.poetry]
name = "sentiment-analysis"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
transformers = "^4.42.4"
torch = "^2.3.1"
pandas = ">=1.3.5"
datasets = "^2.20.0"
pyarrow = "<17.0.0"
urllib3 = "^2.2.2"
qwak-sdk = "^0.5.85"
mlflow = "^2.18.0"

